+++ Editor/Steps/StepPipeline.cs
@@ -18,6 +18,7 @@ namespace Nomnom.UnityProjectPatcher.Editor.Steps {
             new AssetRipperStep(),
             new CopyGamePluginsStep(), // recompile
             new CopyExplicitScriptFolderStep(), // restarts
+            new EnableUnsafeCodeStep(), // recompiles
             new CopyProjectSettingsStep(allowUnsafeCode: true), // restart
             new GuidRemapperStep(),
             new CopyAssetRipperExportToProjectStep(), // restarts