+++ Editor/Steps/StepsExecutor.cs
@@ -113,6 +113,11 @@ namespace Nomnom.UnityProjectPatcher.Editor.Steps {
                             index++;
                             SaveProgress(false);
                             
+                            if (PatcherUtility.LockedAssemblies) {
+                                PatcherUtility.LockedAssemblies = false;
+                                EditorApplication.UnlockReloadAssemblies();
+                            }
+                            
 #if UNITY_2020_3_OR_NEWER
                             CompilationPipeline.RequestScriptCompilation(RequestScriptCompilationOptions.CleanBuildCache);
 #else